support multi-arch images for windows via ctr
